refactor aws string interpolation to its own class
